\name{braidCompile.braidppt}
\alias{braidCompile.braidppt}
\title{Compiles braid object to ppt.}
\usage{
  \method{braidCompile}{braidppt} (b,
    fileOuter = b$fileInner, ...)
}
\arguments{
  \item{b}{A braid object}

  \item{fileOuter}{Location of R script file}

  \item{...}{not used}
}
\description{
  Sources the code created with \code{braidppt}.
}

