\name{braidCompilePPT}
\alias{braidCompilePPT}
\title{Compiles braid object to ppt.}
\usage{
  braidCompilePPT(b = NULL, fileOuter = b$fileInner, ...)
}
\arguments{
  \item{b}{A braid object}

  \item{fileOuter}{Location of R script file}

  \item{...}{not used}
}
\description{
  Sources the code created with \code{braidppt}.
}

