\name{braidpptClose}
\alias{braidpptClose}
\title{Closes ppt.}
\usage{
  braidpptClose(braid)
}
\arguments{
  \item{braid}{Braid object}
}
\description{
  Closes ppt.
}
\seealso{
  Other braidPPT: \code{\link{braidpptInsertImage}},
  \code{\link{braidpptNew}},
  \code{\link{braidpptNewSlide}},
  \code{\link{braidpptSave}}
}

