\name{braidpptInsertImage}
\alias{braidpptInsertImage}
\title{Inserts image on slide.}
\usage{
  braidpptInsertImage(braid, file = NULL, size = NULL)
}
\arguments{
  \item{braid}{Braid object}

  \item{file}{Filename of image to attach}

  \item{size}{Size of image}
}
\description{
  Inserts image on slide.
}
\seealso{
  Other braidPPT: \code{\link{braidpptClose}},
  \code{\link{braidpptNew}},
  \code{\link{braidpptNewSlide}},
  \code{\link{braidpptSave}}
}

