\name{braidpptSave}
\alias{braidpptSave}
\title{Saves ppt.}
\usage{
  braidpptSave(braid, filename)
}
\arguments{
  \item{filename}{Character string: Name of file to save.}

  \item{braid}{Braid object}
}
\description{
  Saves ppt.
}
\seealso{
  Other braidPPT: \code{\link{braidpptClose}},
  \code{\link{braidpptInsertImage}},
  \code{\link{braidpptNew}},
  \code{\link{braidpptNewSlide}},
  \code{\link{braidpptPlot}}
}

