\docType{package}
\name{package-braidppt}
\alias{"package-braidppt"}
\alias{braidppt}
\alias{package-braidppt}
\title{Extension to braid that deals with PowerPoint output}
\description{
  An extension to the braid report writing system that
  creates PowerPoint output from R.
}
\details{
  This package extends braid to create PowerPoint output
  from R.

  Create, save and close a presentation: \itemize{
  \item{\code{\link{pptNew}}} \item{\code{\link{pptSave}}}
  \item{\code{\link{pptClose}}} }

  Add new slides and insert images: \itemize{
  \item{\code{\link{pptNewSlide}}}
  \item{\code{\link{pptInsertImage}}} }
}
\author{
  Andrie de Vries \email{andrie.de.vries@pentalibra.com}
}
\keyword{package}

