\name{pptClose}
\alias{pptClose}
\title{Closes open PowerPoint presentation.}
\usage{
  pptClose(ppt)
}
\arguments{
  \item{ppt}{Pointer to ppt object, generated by
  \code{\link{pptNew}}}
}
\description{
  Closes open PowerPoint presentation.
}
\seealso{
  Other PowerPoint: \code{\link{pptNew}},
  \code{\link{pptNewSlide}}, \code{\link{pptOpen}},
  \code{\link{pptSave}}
}

