\name{pptInsertTextbox}
\alias{pptInsertTextbox}
\title{Insert textbox on current slide.}
\usage{
  pptInsertTextbox(ppt, text, position = c(100, 100, 150,
  50), fontControl = list(Color = 1, Bold = FALSE, Italic =
  FALSE, Underline = FALSE), paraControl = list(Alignment =
  2))
}
\arguments{
  \item{position}{Numeric vector of length 4 indicating
  c(top, left, width, height)}

  \item{fontControl}{List of Properties and values that
  controls the font.  Passed to
  \code{ppt.Current.Slide.Shapes.Item(n).TextFrame.Tex}}

  \item{paraControl}{List of Properties and values that
  controls ParagraphFormat.  The alignment property takes
  integer values (e.g. 1=left, 2=centre, 3=right).}

  \item{ppt}{Pointer to ppt object, generated by
  \code{\link{pptNew}}}

  \item{text}{Slide text}
}
\description{
  Insert textbox on current slide.
}

