\name{pptNewSlide}
\alias{pptNewSlide}
\title{Adds new slide to existing PowerPoint object and adds text and plot.}
\usage{
  pptNewSlide(ppt, title = NULL, text = NULL, subtitle =
  NULL, file = NULL, size = c(0.1, 0.1, 0.9, 0.9))
}
\arguments{
  \item{ppt}{Pointer to ppt object, generated by
  \code{\link{pptNew}}}

  \item{title}{Slide title}

  \item{text}{Slide text}

  \item{subtitle}{Slide subtitle - only added to Title
  slide}

  \item{file}{Location and file name of a plot}

  \item{size}{A numeric vector of size 4, indicating c(x1,
  y1, x2, y2) expressed as percentage of page size}
}
\value{
  A pointer to a ppt object
}
\description{
  Adds new slide to existing PowerPoint object and adds
  text and plot.
}
\examples{
if(isCOMsupported(method="RDCOMClient")){
ppt <- pptNew()
ppt <- pptNewSlide(ppt) # This should be blank
ppt <- pptNewSlide(ppt, "Title", subtitle="Subtitle")
ppt <- pptNewSlide(ppt, "Title only")
ppt <- pptNewSlide(ppt, "Title with text", "Level1\\n\\tLevel 2 item 1\\n\\tLevel 2 item 2\\n\\t\\tLevel 3\\nBack to level 1")
rm(ppt)
}
}
\seealso{
  Other PowerPoint: \code{\link{pptClose}},
  \code{\link{pptNew}}, \code{\link{pptSave}}
}

