\name{pptSave}
\alias{pptSave}
\title{Saves PowerPoint presentation.}
\usage{
  pptSave(ppt, file)
}
\arguments{
  \item{ppt}{Pointer to ppt object, generated by
  \code{\link{pptNew}}}

  \item{file}{Name of file}
}
\description{
  Saves PowerPoint presentation.
}
\seealso{
  Other PowerPoint: \code{\link{pptClose}},
  \code{\link{pptNew}}, \code{\link{pptNewSlide}},
  \code{\link{pptOpen}}
}

