\name{braidpptNew}
\alias{braidpptNew}
\title{Initialises braidppt.}
\usage{
  braidpptNew(braid, template = NULL)
}
\arguments{
  \item{braid}{Braid object}

  \item{template}{Character string: location and file name
  of template to apply}
}
\description{
  Initialises braidppt.
}
\seealso{
  Other braidPPT: \code{\link{braidpptClose}},
  \code{\link{braidpptInsertImage}},
  \code{\link{braidpptNewSlide}},
  \code{\link{braidpptSave}}
}

