\name{braidpptNewSlide}
\alias{braidpptNewSlide}
\title{Adds new slide.}
\usage{
  braidpptNewSlide(braid, title = NULL, text = NULL,
  subtitle = NULL, filename = NULL, size = NULL)
}
\arguments{
  \item{title}{Slide title}

  \item{text}{Slide text}

  \item{subtitle}{Slide subtitle}

  \item{filename}{Filename of image to attach}

  \item{size}{Size of image}

  \item{braid}{Braid object}
}
\description{
  Adds new slide.
}
\seealso{
  Other braidPPT: \code{\link{braidpptClose}},
  \code{\link{braidpptInsertImage}},
  \code{\link{braidpptNew}}, \code{\link{braidpptPlot}},
  \code{\link{braidpptSave}}
}

