\name{braidpptSave}
\alias{braidpptSave}
\title{Saves ppt.}
\usage{
  braidpptSave(braid, filename, path = braid$pathLatex)
}
\arguments{
  \item{filename}{Character string: Name of PowerPoint file
  to save.}

  \item{path}{file path to filename}

  \item{braid}{Braid object}
}
\description{
  Saves ppt.
}
\seealso{
  Other braidPPT: \code{\link{braidpptClose}},
  \code{\link{braidpptInsertImage}},
  \code{\link{braidpptNew}},
  \code{\link{braidpptNewSlide}},
  \code{\link{braidpptPlot}}
}

